EES-2107 - move bootstrapping process into own utility class.  Remove erroneous users from default set of bootstrapping BAU users
