+++ src/GovUk.Education.ExploreEducationStatistics.Admin/appsettings.KeycloakBootstrapUsers.json
@@ -1,5 +1,5 @@
 {
   "BootstrapUsers": {
-    "BAU": "ees-bau1@education.gov.uk,ees-bau2@education.gov.uk,ees-bau3@education.gov.uk"
+    "BAU": "ees-bau1@education.gov.uk,ees-bau2@education.gov.uk"
   }
 }
